bidcruit/requirement1.txt
apturl==0.5.2
asgiref==3.3.1
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
blis==0.7.4
bootstrap4==0.1.0
Brlapi==0.7.0
catalogue==2.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
commonregex==1.5.4
cryptography==2.8
cupshelpers==1.0
cymem==2.0.5
datefinder==0.7.1
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.4
django-bootstrap4==2.3.1
django-crispy-forms==1.8.1
django-user-agents==0.4.0
djangorestframework==3.12.4
docopt==0.6.2
docx2txt==0.8
duplicity==0.8.12.0
elasticsearch==7.12.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
httplib2==0.14.0
idna==2.9
Jinja2==2.11.3
joblib==1.0.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
murmurhash==1.0.5
nameparser==1.0.6
netifaces==0.10.4
nltk==3.5
numpy==1.20.1
oauthlib==3.1.0
olefile==0.46
packaging==20.9
paramiko==2.6.0
pathy==0.4.0
pdfminer==20191016
pexpect==4.6.0
Pillow==7.0.0
pipreqs==0.4.10
preshed==3.0.5
protobuf==3.6.1
psycopg2-binary==2.8.6
pycairo==1.16.2
pycryptodome==3.10.1
pycups==1.9.73
pydantic==1.7.3
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyotp==2.6.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
regex==2021.3.17
reportlab==3.5.34
requests==2.23.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
smart-open==3.0.0
soupsieve==2.2.1
spacy==3.0.5
spacy-legacy==3.0.1
sqlparse==0.3.1
srsly==2.4.0
stripe==2.42.0
systemd-python==234
thinc==8.0.2
tqdm==4.59.0
typer==0.3.2
ua-parser==0.10.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.7
user-agents==2.2.0
wadllib==1.3.3
wasabi==0.8.2
xkit==0.0.0
yarg==0.1.9
